changed how width and height is recomputed in metadata
