+++ raster_loader/io.py
@@ -1,9 +1,7 @@
-from functools import partial
 import time
+from functools import partial
 import json
 import sys
-import math
-from itertools import islice
 from typing import Iterable
 from typing import Callable
 from typing import List
@@ -13,7 +11,6 @@ from affine import Affine
 import numpy as np
 import pandas as pd
 import pyproj
-import functools
 import shapely
 
 try:
@@ -44,95 +41,12 @@ except ImportError:  # pragma: no cover
 else:
     _has_bigquery = True
 
-from raster_loader.utils import ask_yes_no_question
+from raster_loader.utils import ask_yes_no_question, batched
+from raster_loader.geo import (
+    raster_bounds,
+)
 
 should_swap = {"=": sys.byteorder != "little", "<": False, ">": True, "|": False}
-
-
-def batched(iterable, n):
-    "Batch data into tuples of length n. The last batch may be shorter."
-    # batched('ABCDEFG', 3) --> ABC DEF G
-    if n < 1:  # pragma: no cover
-        raise ValueError("n must be at least one")
-    it = iter(iterable)
-    while batch := tuple(islice(it, n)):  # noqa
-        yield batch
-
-
-def coord_range(start_x, start_y, end_x, end_y, num_subdivisions):
-    num_subdivisions = max(num_subdivisions, 1)
-    return [
-        [
-            start_x + (end_x - start_x) * i / num_subdivisions,
-            start_y + (end_y - start_y) * i / num_subdivisions,
-        ]
-        for i in range(0, num_subdivisions + 1)
-    ]
-
-
-def norm_lon(x):
-    return x - 360.0 if x > 180.0 else x + 360.0 if x <= -180.0 else x
-
-
-def norm_coords(coords):
-    return [[norm_lon(point[0]), point[1]] for point in coords]
-
-
-def polygon_geography(rings, format, normalize_coords):
-    if normalize_coords:
-        rings = [norm_coords(coords) for coords in rings]
-
-    if format == "wkt":
-        return polygon_wkt(rings)
-    elif format == "geojson":
-        return polygon_geojson(rings)
-    else:
-        raise ValueError(f"Invalid geography format {format}")
-
-
-def polygon_wkt(rings):
-    return (
-        "POLYGON("
-        + ",".join(
-            [
-                "("
-                + ",".join(
-                    [" ".join([str(coord) for coord in point]) for point in coords]
-                )
-                + ")"
-                for coords in rings
-            ]
-        )
-        + ")"
-    )
-
-
-def polygon_geojson(rings):
-    return json.dumps({"type": "Polygon", "coordinates": rings})
-
-
-def coords_to_geography(coords, format, whole_earth):
-    # remove too-close coordinates cause they cause errors
-    # in BigQuery's ST_GEOGFROMGEOJSON
-    def are_too_close(point1, point2):
-        return (
-            math.fabs(point1[0] - point2[0]) <= 1e-13
-            and math.fabs(point1[1] - point2[1]) <= 1e-13
-        )
-
-    def filter_near_points(coords, point):
-        previous = None if not coords else coords[-1]
-        if not previous or not are_too_close(previous, point):
-            coords.append(point)
-        return coords
-
-    coords = functools.reduce(filter_near_points, coords, [])
-
-    # now let's make sure the initial and final points are exactly the same
-    if coords[0] != coords[-1]:
-        # replace the last point; never mind, it must be very close
-        coords[-1] = coords[0]
-    return polygon_geography([coords], format, not whole_earth)
 
 
 def band_field_name(custom_name: str, band: int) -> str:
@@ -219,40 +133,6 @@ def table_columns(bands: List[str]) -> List[Tuple[str, str, str]]:
     columns = [("block", "INTEGER", "REQUIRED"), ("metadata", "STRING", "NULLABLE")]
     columns += [(band_name, "BYTES", "NULLABLE") for band_name in bands]
     return columns
-
-
-def raster_bounds(raster_dataset, transformer, format):
-    raster_dataset.transform,
-    min_x = 0
-    min_y = 0
-    max_x = raster_dataset.width
-    max_y = raster_dataset.height
-
-    x_subdivisions = math.ceil((max_x - min_x) / 64.0)
-    y_subdivisions = math.ceil((max_y - min_y) / 64.0)
-    pixel_coords = (
-        # SW -> SE
-        coord_range(min_x, max_y, max_x, max_y, x_subdivisions)
-        # SE -> NE
-        + coord_range(max_x, max_y, max_x, min_y, y_subdivisions)
-        # NE -> NW
-        + coord_range(max_x, min_y, min_x, min_y, x_subdivisions)
-        # NW -> SW
-        + coord_range(min_x, min_y, min_x, max_y, y_subdivisions)
-    )
-    coords = [
-        transformer.transform(*(raster_dataset.transform * (x, y)))
-        for x, y in pixel_coords
-    ]
-    lon_NW, _ = transformer.transform(*(raster_dataset.transform * (min_x, min_y)))
-    lon_NE, _ = transformer.transform(*(raster_dataset.transform * (max_x, min_y)))
-    lon_SW, _ = transformer.transform(*(raster_dataset.transform * (min_x, max_y)))
-    lon_SE, _ = transformer.transform(*(raster_dataset.transform * (max_x, max_y)))
-    whole_earth = (
-        math.fabs(lon_NW - lon_NE) >= 360.0 and math.fabs(lon_SW - lon_SE) >= 360
-    )
-
-    return coords_to_geography(coords, format, whole_earth)
 
 
 def rasterio_metadata(
@@ -899,8 +779,14 @@ def update_metadata(metadata, old_metadata):
     )
     metadata["num_blocks"] += old_metadata["num_blocks"]
     metadata["num_pixels"] += old_metadata["num_pixels"]
-    metadata["width"] = None
-    metadata["height"] = None
+    w, s, _ = quadbin.utils.point_to_tile(
+        metadata["bounds"][0], metadata["bounds"][1], metadata["resolution"]
+    )
+    e, n, _ = quadbin.utils.point_to_tile(
+        metadata["bounds"][2], metadata["bounds"][3], metadata["resolution"]
+    )
+    metadata["height"] = (s - n) * metadata["block_height"]
+    metadata["width"] = (e - w) * metadata["block_width"]
 
 
 def get_metadata(project_id, dataset_id, table_id, client=None):
