+++ raster_loader/utils.py
@@ -1,3 +1,6 @@
+from itertools import islice
+
+
 def ask_yes_no_question(question: str) -> bool:
     """Ask a yes or no question and return True or False."""
     yes_choices = ["yes", "y"]
@@ -12,3 +15,13 @@ def ask_yes_no_question(question: str) -> bool:
         else:  # pragma: no cover
             print("Type yes or no")
             continue
+
+
+def batched(iterable, n):
+    "Batch data into tuples of length n. The last batch may be shorter."
+    # batched('ABCDEFG', 3) --> ABC DEF G
+    if n < 1:  # pragma: no cover
+        raise ValueError("n must be at least one")
+    it = iter(iterable)
+    while batch := tuple(islice(it, n)):  # noqa
+        yield batch
